+++ packages/use-intl/src/index.tsx
@@ -3,6 +3,7 @@ export {default as IntlMessages} from './IntlMessages';
 export {default as useTranslations} from './useTranslations';
 export {default as TranslationValues} from './TranslationValues';
 export {default as useIntl} from './useIntl';
+export {default as useNow} from './useNow';
 export {default as Formats} from './Formats';
 export {default as DateTimeFormatOptions} from './DateTimeFormatOptions';
 export {default as IntlError, IntlErrorCode} from './IntlError';